+++ src/Food/Food.cpp
@@ -1,10 +1,10 @@
 #include "Food.hpp"
 
-Food::Food(int x, int y, int width, int height) : Item(x, y, width, height, TYPE_FOOD){
+Food::Food(double x, double y, double width, double height) : Item(x, y, width, height, TYPE_FOOD){
 }
 
 void Food::tick(int delay){
-	int y = this->getY();
+	double y = this->getY();
 	if (!(this->isAtBottom)){
 		y = y + 1;
 		this->setY(y);
